bdg-eng-ist-ondc-qa-main/bdg-eng-ist-ondc-qa-main/requirements.txt
# HTTP & networking
requests>=2.33.0
urllib3>=2.6.3

# CLI framework
click>=8.1.7

# Configuration
PyYAML>=6.0.1

# Cryptographic utilities
cryptography>=46.0.5
PyNaCl>=1.6.2

# JSON schema validation
jsonschema>=4.23.0

# Report generation
beautifulsoup4>=4.12.0
openpyxl>=3.1.5
defusedxml>=0.7.1    # Recommended by openpyxl for XML attack protection
pandas>=2.2.0
Jinja2>=3.1.6

# Development tools
debugpy>=1.8.0
